Replace old transition relation computation with new one
